requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
sentence-transformers
torch
faiss-cpu
pypdf
python-multipart
streamlit
requests
pytest
pytest-asyncio
httpx
